Keychron refactor.
