+++ keyboards/keychron/q3/rev_0120/keymaps/keychron_default/rules.mk
@@ -1,0 +1,2 @@
+VPATH += keyboards/keychron/common
+SRC += keychron_common.c
