+++ keyboards/keychron/q3/rev_0120/keymaps/keychron_via/rules.mk
@@ -1,0 +1,4 @@
+VPATH += keyboards/keychron/common
+SRC += keychron_common.c
+
+VIA_ENABLE = yes
